Add parsing for attributes and attibutes on operations.  Add IR representation for attributes on operations.  Split Operation out from OperationInst so it can be shared with OperationStmt one day.
PiperOrigin-RevId: 203325366

+++ include/mlir/IR/Attributes.h
@@ -83,7 +83,7 @@ class IntegerAttr : public Attribute {
 public:
   static IntegerAttr *get(int64_t value, MLIRContext *context);
 
-  unsigned getValue() const {
+  int64_t getValue() const {
     return value;
   }
